eclass/racket.eclass updated to use installation scope
racket.eclass cleanup

moving everything to installation scope removes the need for any changes
to baselayout so also removed that as a dependency

racket.eclass switch to use compile-directory-zos

works better than raco make because we can read the info files and avoid
issues with files that aren't supposed to be compiled according to the
info.rkt files, still need to investigate whether this approach compiles
in parallel or not

racket.eclass fix pkg install during compile

also fixed naming of a number of functions and removed the portage user
dependency

racket.eclass remove EPREFIX for paths used in D

racket.eclass update authors

racket.eclass fix P_RACKET_DIR quoting

racket.eclass remove CONFIG_PROTECT

Since all ebuilds using the racket eclass have to be rebuilt when racket
is upgraded we don't need to protect links.rktd and pkgs.rktd, the only
time when someone could get into trouble is if they remerge the same
version of racket, however if they do that they are probably a developer
and thus can be expected to find this note (or similar).

racket.eclass #:skip-doc-sources? #t in compile

this avoids issues with cases where documentation expects a module to
already be installed under the expected name when that module shares the
folder that must be named correctly with the working directory of the
git repository or the name of the source tarball both of which are
technically arbitrary